% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_graph_utils_create_subgraph.R
\name{createSubgraph}
\alias{createSubgraph}
\title{Create Subgraph}
\usage{
createSubgraph(g, node, order, mode = c("all", "in", "out"))
}
\arguments{
\item{g}{Igraph}

\item{node}{Character}

\item{order}{Integer}

\item{mode}{Character}
}
\value{
Igraph
}
\description{
Create Subgraph
}
