% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dependencies}
\alias{dependencies}
\title{Dependencies}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 703 rows and 5 columns.
}
\usage{
dependencies
}
\description{
Dependencies
}
\keyword{datasets}
