% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_data.R
\name{FunctionData}
\alias{FunctionData}
\title{FunctionData}
\description{
R6 class wrapper for data returned by \code{functiondepends} functions
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Character path of repository}

\item{\code{envir}}{Environment to store parsed functions in}

\item{\code{functions}}{Data.frame with functions found in \code{path}}

\item{\code{dependencies}}{Data.frame serving as edge list of function dependencies}

\item{\code{graph, }}{Igraph object}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{graph, }}{Igraph object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FunctionData$new()}}
\item \href{#method-is_initialized}{\code{FunctionData$is_initialized()}}
\item \href{#method-reset}{\code{FunctionData$reset()}}
\item \href{#method-download_github}{\code{FunctionData$download_github()}}
\item \href{#method-set_path}{\code{FunctionData$set_path()}}
\item \href{#method-get_path}{\code{FunctionData$get_path()}}
\item \href{#method-set_envir}{\code{FunctionData$set_envir()}}
\item \href{#method-get_envir}{\code{FunctionData$get_envir()}}
\item \href{#method-find_functions}{\code{FunctionData$find_functions()}}
\item \href{#method-get_functions}{\code{FunctionData$get_functions()}}
\item \href{#method-find_dependencies}{\code{FunctionData$find_dependencies()}}
\item \href{#method-get_dependencies}{\code{FunctionData$get_dependencies()}}
\item \href{#method-make_graph}{\code{FunctionData$make_graph()}}
\item \href{#method-get_graph}{\code{FunctionData$get_graph()}}
\item \href{#method-plot}{\code{FunctionData$plot()}}
\item \href{#method-plot_distribution}{\code{FunctionData$plot_distribution()}}
\item \href{#method-clone}{\code{FunctionData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize path and create new environment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionData$new()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Character path to search source R files in}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_initialized"></a>}}
\if{latex}{\out{\hypertarget{method-is_initialized}{}}}
\subsection{Method \code{is_initialized()}}{
Check if fields are initialized
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionData$is_initialized()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
Reset fields to NULL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionData$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_github"></a>}}
\if{latex}{\out{\hypertarget{method-download_github}{}}}
\subsection{Method \code{download_github()}}{
Download repository from GitHub
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionData$download_github(repo, branch)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{repo}}{Character, name of repository in format Name/Repo}

\item{\code{branch}}{Character, name of branch}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical, indicates whether download succeeded
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_path"></a>}}
\if{latex}{\out{\hypertarget{method-set_path}{}}}
\subsection{Method \code{set_path()}}{
Set \code{path} field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionData$set_path(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Character path to search source R files in}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_path"></a>}}
\if{latex}{\out{\hypertarget{method-get_path}{}}}
\subsection{Method \code{get_path()}}{
Getter for \code{path} field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionData$get_path()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_envir"></a>}}
\if{latex}{\out{\hypertarget{method-set_envir}{}}}
\subsection{Method \code{set_envir()}}{
Set \code{envir} field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionData$set_envir(envir = new.env())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{envir}}{Environment, default is \code{new.env()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_envir"></a>}}
\if{latex}{\out{\hypertarget{method-get_envir}{}}}
\subsection{Method \code{get_envir()}}{
Getter for \code{envir} field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionData$get_envir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find_functions"></a>}}
\if{latex}{\out{\hypertarget{method-find_functions}{}}}
\subsection{Method \code{find_functions()}}{
Search for functions in \code{path}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionData$find_functions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_functions"></a>}}
\if{latex}{\out{\hypertarget{method-get_functions}{}}}
\subsection{Method \code{get_functions()}}{
Getter for \code{functions} field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionData$get_functions()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find_dependencies"></a>}}
\if{latex}{\out{\hypertarget{method-find_dependencies}{}}}
\subsection{Method \code{find_dependencies()}}{
Find dependencies of found functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionData$find_dependencies(in_envir = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{in_envir}}{Logical, to be passed to \code{functiondepends::find_dependencies}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dependencies"></a>}}
\if{latex}{\out{\hypertarget{method-get_dependencies}{}}}
\subsection{Method \code{get_dependencies()}}{
Getter for \code{dependencies} field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionData$get_dependencies()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_graph"></a>}}
\if{latex}{\out{\hypertarget{method-make_graph}{}}}
\subsection{Method \code{make_graph()}}{
Make graph
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionData$make_graph()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_graph"></a>}}
\if{latex}{\out{\hypertarget{method-get_graph}{}}}
\subsection{Method \code{get_graph()}}{
Get \code{graph}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionData$get_graph(node = NULL, mode = NULL, order = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node}}{Character, name of node to select}

\item{\code{mode}}{Character, mode of neighborhood}

\item{\code{order}}{Integer, order of neighborhood}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plot network
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionData$plot(node = NULL, mode = NULL, order = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node}}{Character, name of node to select}

\item{\code{mode}}{Character, mode of neighborhood}

\item{\code{order}}{Integer, order of neighborhood}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_distribution"></a>}}
\if{latex}{\out{\hypertarget{method-plot_distribution}{}}}
\subsection{Method \code{plot_distribution()}}{
Plot network
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionData$plot_distribution(node = NULL, mode = "all", order = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node}}{Character, name of node to select}

\item{\code{mode}}{Character, mode of neighborhood}

\item{\code{order}}{Integer, order of neighborhood}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FunctionData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
