% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{functions}
\alias{functions}
\title{Functions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 3 columns.
}
\usage{
functions
}
\description{
Functions
}
\keyword{datasets}
